+++ CollAction/Models/ProjectViewModels/EditProjectViewModel.cs
@@ -62,13 +62,13 @@ namespace CollAction.Models
         [Required]
         [Display(Name = "Start date")]
         [DisplayFormat(DataFormatString = "{0:dd/MM/yyyy}", ApplyFormatInEditMode = true)]
-        [DateWithinMonths(12, ErrorMessage = "Please ensure your project starts within the next 12 months.")]
+        [WithinMonthsAfterToday(12, ErrorMessage = "Please ensure your project starts within the next 12 months.")]
         public DateTime Start { get; set; }
 
         [Required]
         [Display(Name = "Deadline")]
         [DisplayFormat(DataFormatString = "{0:dd/MM/yyyy}", ApplyFormatInEditMode = true)]
-        [DateWithinMonths(12, OfDateProperty = "Start", ErrorMessage = "The deadline must be within a year of the start date.")]
+        [WithinMonthsAfterDateProperty(12, "Start", ErrorMessage = "The deadline must be within a year of the start date.")]
         public DateTime End { get; set; }
 
         [Display(Name = "Banner image", Prompt = "1024x768px JPEG, GIF, PNG, BMP")]
